+++ desktop/src-tauri/src/commands/mod.rs
@@ -46,6 +46,7 @@ mod social;
 mod team_snapshot;
 mod teams;
 mod updater;
+mod window_actions;
 mod window_vibrancy;
 mod workflows;
 mod workspace;
@@ -93,6 +94,7 @@ pub use social::*;
 pub use team_snapshot::*;
 pub use teams::*;
 pub use updater::*;
+pub use window_actions::*;
 pub use window_vibrancy::*;
 pub use workflows::*;
 pub use workspace::*;
